Doc: Convert doc config to use standard Qt module make targets
With qmllive as a submodule of a module that uses the stardard
Qt documentation targets (prepare_docs, generate_docs, etc.),
the build was failing. Fix this by adopting the same
documentation targets as a Qt module.

This allows doc builds of qmllive as part of a larger project.
Also, because there is now a prepare_docs step that generates
.index files, this enables proper cross-module linking via doc
dependendencies.

Task-number: AUTOSUITE-8
Change-Id: I553487eeea80ce288c417cf42f7052e5c1318c67
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>