Allow to override default QSettings location
Inspired by Qt Creator's IDE_SETTINGSVARIANT.

Useful when distributing QmlLive as part of a development kit, to
separate settings from other QmlLive installations installed
on the same workstation.

This is achieved by setting QCoreApplication::organizationName.
This property is dedicated for use by QSettings when it is
constructed using the default constructor.

Change-Id: I6f0d273fbbd89d23289618b0d3a061c1f797a585
Reviewed-by: Juergen Bocklage-Ryannel <juergen.bocklage-ryannel@pelagicore.com>