Fix documentation build on Clang-enabled QDoc
When building documentation with QDoc from Qt 5.11 or later, C++
sources are parsed by Clang. Clang expects a complete set of
include paths for resolving the types in the documented API, as
well as a module header to build a precompiled header from.

To do this, convert the /doc directory to a proper subproject
and add suitable dependencies to get the right include paths
from qmake. As QmlLive is not a Qt module, docs must also
maintain a 'dummy' module header that is passed to Clang.

Fix also remaining documentation warnings, most importantly,
the syntax of \fn commands.

Change-Id: I2e912dbc509e2bd40691463e4608081372b5d6a2
Reviewed-by: Kavindra Palaraja <kpalaraja@luxoft.com>
Reviewed-by: Svetlana Abramenkova <sabramenkova@luxoft.com>